[fix] one post has multiple categories
