+++ database/seeders/DatabaseSeeder.php
@@ -16,7 +16,8 @@ class DatabaseSeeder extends Seeder
         $this->call(UsersTableSeeder::class);
         $this->call(RolesTableSeeder::class);
         $this->call(PermissionsTableSeeder::class);
-        $this->call(PostCategoriesTableSeeder::class);
+        $this->call(CategoriesTableSeeder::class);
         $this->call(PostsTableSeeder::class);
+        $this->call(PostCategoriesTableSeeder::class);
     }
 }
